Fixed bug caused by incorrect ordering
This was a weird bug! If you don't order a django queryset, you get unpredictable behaviour.
In this instance, the un-ordered code worked perfectly in my IDE (pycharm w/ debugger), but failed on production (nginx + gunicorn + postgress). On the server, requesting slots[0] *Aways* gave the last item in the list, as did slots[1]. Weird!